Migrate package builds to tsdown
