+++ .gitignore
@@ -29,7 +29,7 @@ node_modules/
 .wireit
 .eslintcache
 .tmp
-tsup.config.bundled_*.mjs
+tsdown.config.bundled_*.mjs
 build.utils.d.ts
 worker-configuration.d.ts
 /.env
