+++ package.json
@@ -16,7 +16,7 @@
     "clean:integration": "node ./integration/helpers/cleanup.mjs",
     "docs": "pnpm run docs:typedoc && pnpm run docs:jsdoc",
     "docs:typedoc": "typedoc",
-    "docs:jsdoc": "node --experimental-strip-types scripts/docs.ts",
+    "docs:jsdoc": "node scripts/docs.ts",
     "experimental:version": "node ./scripts/experimental.ts version",
     "experimental:publish": "node ./scripts/experimental.ts publish",
     "format": "prettier --ignore-path .prettierignore --write .",
@@ -26,8 +26,8 @@
     "previews:branch": "node ./scripts/previews/branch.ts",
     "previews:pr": "node ./scripts/previews/pr.ts",
     "release-comments": "node scripts/release-comments.ts",
-    "test": "node --experimental-vm-modules ./node_modules/jest/bin/jest.js",
-    "test:inspect": "node --experimental-vm-modules --inspect-brk ./node_modules/jest/bin/jest.js",
+    "test": "node --experimental-vm-modules --no-warnings=ExperimentalWarning ./node_modules/jest/bin/jest.js",
+    "test:inspect": "node --experimental-vm-modules --no-warnings=ExperimentalWarning --inspect-brk ./node_modules/jest/bin/jest.js",
     "typegen": "pnpm run --recursive --parallel typegen",
     "typecheck": "pnpm run --recursive --parallel typecheck",
     "test:integration:run": "pnpm playwright:integration",
