+++ packages/react-router-express/package.json
@@ -28,7 +28,7 @@
   },
   "wireit": {
     "build": {
-      "command": "tsup",
+      "command": "tsdown --config-loader unrun",
       "files": [
         "../../pnpm-workspace.yaml",
         "*.ts",
@@ -50,7 +50,7 @@
     "express": "^4.19.2",
     "node-mocks-http": "^1.17.2",
     "supertest": "^6.3.3",
-    "tsup": "catalog:",
+    "tsdown": "catalog:",
     "typescript": "catalog:",
     "wireit": "catalog:"
   },
