+++ packages/react-router/lib/rsc/server.rsc.ts
@@ -48,7 +48,7 @@ import {
   UNSAFE_WithComponentProps,
   UNSAFE_WithHydrateFallbackProps,
   UNSAFE_WithErrorBoundaryProps,
-  // @ts-ignore There are no types before the tsup build when used internally, so
+  // @ts-ignore There are no types before the package build when used internally, so
   // we need to cast. If we add an alias for 'internal/react-server-client' to our
   // TSConfig, it breaks the Parcel build.
 } from "react-router/internal/react-server-client";
